backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pymongo>=4.8.0
python-dotenv>=1.0.1
google-generativeai>=0.8.3
dnspython>=2.6.1
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
